renamed 'validations_widget.py' to 'report_page.py'
